Extrai metodo get_regras para um módulo
