+++ dominio/suamesa.py
@@ -1,0 +1,49 @@
+from functools import lru_cache
+
+from django.conf import settings
+
+from .db_connectors import run_query
+
+
+QUERY_REGRAS = """
+    SELECT r.classe_documento
+    FROM {namespace}.atualizacao_pj_pacote pct
+    JOIN {namespace}.{regras_table} r
+    ON r.cod_atribuicao = pct.cod_pct
+    WHERE pct.id_orgao = {orgao_id}
+    """
+
+
+@lru_cache()
+def get_regras(orgao_id, tipo='investigacao'):
+    """Busca as regras de negócio relativas a investigação ou processo,
+    para um dado órgão.
+
+    Arguments:
+        orgao_id {integer} -- ID do órgão a ser procurado.
+
+    Keyword Arguments:
+        tipo {str} -- Tipo de regra a ser procurada.
+        Pode ser 'investigacao' ou 'processo'. (default: {'investigacao'})
+
+    Returns:
+        List[integer] -- Lista de IDs de classes de documentos.
+    """
+    table_switcher = {
+        'investigacao': 'tb_regra_negocio_investigacao',
+        'processo': 'tb_regra_negocio_processo'
+    }
+    regras_table = table_switcher.get(tipo, None)
+
+    if not regras_table:
+        # Melhor fazer de outra forma? Raise error?
+        return None
+
+    query = QUERY_REGRAS.format(
+        orgao_id=orgao_id,
+        regras_table=regras_table,
+        namespace=settings.TABLE_NAMESPACE
+    )
+
+    result = run_query(query)
+    return [row[0] for row in result] if result else []
