+++ requirements.txt
@@ -20,6 +20,5 @@ psycopg2-binary
 pyjwt
 python-decouple
 redis
-ring==0.7.3
 requests
 responses
